Refactor internal packages
