+++ core/src/main/kotlin/dev/skye/daisy/action/MessageDeleter.kt
@@ -1,24 +1,13 @@
-package dev.skye.daisy
+package dev.skye.daisy.action
 
+import dev.skye.daisy.logger
+import dev.skye.daisy.poller.QueuePoller
+import dev.skye.daisy.utility.deletedCounter
 import io.micrometer.core.instrument.MeterRegistry
 import kotlinx.coroutines.future.await
 import software.amazon.awssdk.core.exception.SdkException
 import software.amazon.awssdk.services.sqs.SqsAsyncClient
 import software.amazon.awssdk.services.sqs.model.DeleteMessageRequest
-
-internal sealed class DeleteResult {
-
-    internal data class Failure(val cause: Throwable) : DeleteResult()
-    internal object Success : DeleteResult()
-}
-
-internal interface MessageDeleting {
-
-    suspend fun delete(
-        queueUrl: String,
-        receiptHandle: String
-    ): DeleteResult
-}
 
 internal class MessageDeleter(
     private val client: SqsAsyncClient,
