+++ core/src/main/kotlin/dev/skye/daisy/processor/MessageProcessing.kt
@@ -1,5 +1,6 @@
-package dev.skye.daisy
+package dev.skye.daisy.processor
 
+import dev.skye.daisy.action.PostProcessAction
 import software.amazon.awssdk.services.sqs.model.Message
 
 public interface MessageProcessing {
