+++ core/src/main/kotlin/dev/skye/daisy/work/WorkSampling.kt
@@ -1,0 +1,10 @@
+package dev.skye.daisy.work
+
+import kotlinx.coroutines.channels.Channel
+
+internal interface WorkSampling {
+
+    suspend fun sample(
+        inputs: List<Channel<Work>>
+    ): Work
+}
